+++ EnhanceQoLMythicPlus/Init.lua
@@ -66,6 +66,7 @@ addon.functions.InitDBValue("mythicPlusBRButtonSize", 50)
 addon.functions.InitDBValue("mythicPlusBRTrackerPoint", "CENTER")
 addon.functions.InitDBValue("mythicPlusBRTrackerX", 0)
 addon.functions.InitDBValue("mythicPlusBRTrackerY", 0)
+addon.functions.InitDBValue("mythicPlusBRTrackerShowAnchor", false)
 
 -- Talent Reminder
 addon.functions.InitDBValue("talentReminderEnabled", false)